chore: create a separate file for plugin
