+++ examples/minimal-plugin/index.js
@@ -2,30 +2,7 @@
 
 const Engine = require('@neobbs/core');
 const createRenderer = require('@neobbs/renderer-terminal');
-
-// --- Plugin ---
-
-function examplePlugin(engine) {
-  engine.registerPermission('example:item:read');
-
-  engine.registerService('example', {
-    getMessage() {
-      return 'Hello from the example service.';
-    },
-  });
-
-  engine.registerRoute('example', (session) => {
-    const example = engine.getService('example');
-
-    session.send({ type: 'header', title: 'Minimal Plugin' });
-    session.send(example.getMessage());
-    session.send({ type: 'footer' });
-  });
-}
-
-module.exports = examplePlugin;
-
-// --- Application ---
+const examplePlugin = require('./plugin');
 
 const engine = new Engine();
 engine.use(examplePlugin);
